+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `next_id` now returns an `Id`, which can be dereferenced to a `u64`
+
 ## [0.4.0] - 2025-07-20
 
 ### Changed